remove useless BackendAIOConnection
