+++ src/main/java/com/actiontech/dble/backend/mysql/LoadDataUtil.java
@@ -7,7 +7,7 @@ package com.actiontech.dble.backend.mysql;
 
 import com.actiontech.dble.DbleServer;
 import com.actiontech.dble.backend.BackendConnection;
-import com.actiontech.dble.net.BackendAIOConnection;
+import com.actiontech.dble.backend.mysql.nio.MySQLConnection;
 import com.actiontech.dble.net.mysql.BinaryPacket;
 import com.actiontech.dble.route.RouteResultsetNode;
 import com.actiontech.dble.sqlengine.mpp.LoadData;
@@ -24,7 +24,7 @@ public final class LoadDataUtil {
 
     public static void requestFileDataResponse(byte[] data, BackendConnection conn) {
         byte packId = data[3];
-        BackendAIOConnection backendAIOConnection = (BackendAIOConnection) conn;
+        MySQLConnection c = (MySQLConnection) conn;
         RouteResultsetNode rrn = (RouteResultsetNode) conn.getAttachment();
         LoadData loadData = rrn.getLoadData();
         List<String> loadDataData = loadData.getData();
@@ -36,9 +36,9 @@ public final class LoadDataUtil {
                     byte[] bytes = s.getBytes(CharsetUtil.getJavaCharset(loadData.getCharset()));
                     bos.write(bytes);
                 }
-                packId = writeToBackConnection(packId, new ByteArrayInputStream(bos.toByteArray()), backendAIOConnection);
+                packId = writeToBackConnection(packId, new ByteArrayInputStream(bos.toByteArray()), c);
             } else {
-                packId = writeToBackConnection(packId, new BufferedInputStream(new FileInputStream(loadData.getFileName())), backendAIOConnection);
+                packId = writeToBackConnection(packId, new BufferedInputStream(new FileInputStream(loadData.getFileName())), c);
             }
         } catch (IOException e) {
             throw new RuntimeException(e);
@@ -46,14 +46,14 @@ public final class LoadDataUtil {
             //send empty packet
             byte[] empty = new byte[]{0, 0, 0, 3};
             empty[3] = ++packId;
-            backendAIOConnection.write(empty);
+            c.write(empty);
         }
     }
 
-    public static byte writeToBackConnection(byte packID, InputStream inputStream, BackendAIOConnection backendAIOConnection) throws IOException {
+    public static byte writeToBackConnection(byte packID, InputStream inputStream, MySQLConnection c) throws IOException {
         try {
             int packSize = DbleServer.getInstance().getConfig().getSystem().getBufferPoolChunkSize() - 5;
-            // int packSize = backendAIOConnection.getMaxPacketSize() / 32;
+            // int packSize = c.getMaxPacketSize() / 32;
             //  int packSize=65530;
             byte[] buffer = new byte[packSize];
             int len = -1;
@@ -69,7 +69,7 @@ public final class LoadDataUtil {
                 BinaryPacket packet = new BinaryPacket();
                 packet.setPacketId(++packID);
                 packet.setData(temp);
-                packet.write(backendAIOConnection);
+                packet.write(c);
             }
         } finally {
             inputStream.close();
