+++ src/main/java/com/actiontech/dble/net/AIOConnector.java
@@ -6,6 +6,7 @@
 package com.actiontech.dble.net;
 
 import com.actiontech.dble.DbleServer;
+import com.actiontech.dble.backend.mysql.nio.MySQLConnection;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -16,7 +17,7 @@ import java.util.concurrent.atomic.AtomicLong;
  * @author mycat
  */
 public final class AIOConnector implements SocketConnector,
-        CompletionHandler<Void, BackendAIOConnection> {
+        CompletionHandler<Void, MySQLConnection> {
     private static final Logger LOGGER = LoggerFactory.getLogger(AIOConnector.class);
     private static final ConnectIdGenerator ID_GENERATOR = new ConnectIdGenerator();
 
@@ -25,16 +26,16 @@ public final class AIOConnector implements SocketConnector,
     }
 
     @Override
-    public void completed(Void result, BackendAIOConnection attachment) {
+    public void completed(Void result, MySQLConnection attachment) {
         finishConnect(attachment);
     }
 
     @Override
-    public void failed(Throwable exc, BackendAIOConnection conn) {
+    public void failed(Throwable exc, MySQLConnection conn) {
         conn.onConnectFailed(exc);
     }
 
-    private void finishConnect(BackendAIOConnection c) {
+    private void finishConnect(MySQLConnection c) {
         try {
             if (c.finishConnect()) {
                 c.setId(ID_GENERATOR.getId());
