+++ src/main/java/com/actiontech/dble/net/mysql/CommandPacket.java
@@ -8,7 +8,7 @@ package com.actiontech.dble.net.mysql;
 import com.actiontech.dble.backend.mysql.BufferUtil;
 import com.actiontech.dble.backend.mysql.MySQLMessage;
 import com.actiontech.dble.backend.mysql.StreamUtil;
-import com.actiontech.dble.net.BackendAIOConnection;
+import com.actiontech.dble.backend.mysql.nio.MySQLConnection;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -93,7 +93,7 @@ public class CommandPacket extends MySQLPacket {
     }
 
     @Override
-    public void write(BackendAIOConnection c) {
+    public void write(MySQLConnection c) {
         ByteBuffer buffer = c.allocate();
         try {
             BufferUtil.writeUB3(buffer, calcPacketSize());
