+++ src/main/java/com/actiontech/dble/net/mysql/MySQLPacket.java
@@ -5,7 +5,7 @@
 */
 package com.actiontech.dble.net.mysql;
 
-import com.actiontech.dble.net.BackendAIOConnection;
+import com.actiontech.dble.backend.mysql.nio.MySQLConnection;
 import com.actiontech.dble.net.FrontendConnection;
 
 import java.nio.ByteBuffer;
@@ -181,7 +181,7 @@ public abstract class MySQLPacket {
     /**
      * write to backend connection
      */
-    public void write(BackendAIOConnection c) {
+    public void write(MySQLConnection c) {
         throw new UnsupportedOperationException();
     }
